final_project/requirements.txt
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
PyAudio==0.2.13
pydub==0.25.1
python-dotenv==1.0.0
requests==2.29.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
rgbmatrix==0.0.1
scipy==1.6.0
numpy==1.24.3
